Random Integer Calculation and Ordering
The following program creates an array of 25 randomly generated numbers. Sum, average, median, and mode are calculated for the set of randomly generated numbers. The array is sorted, then exported to a text file. The text file display the randomly generated numbers in order. The text file is then read and the values of the file displayed on the console.